% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codons}
\alias{codons}
\title{Codon Translation Table}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://r4bds.github.io/lab08.html}
}
\usage{
codons
}
\description{
A list of codons and the corresponding amino acid they encode for. Stop codons are represented as a "_" character.
}
\keyword{datasets}
